[fix] Refactored old editor
