+++ src/EasyEPlanner.csproj
@@ -115,7 +115,13 @@
   <ItemGroup>
     <Compile Include="Configuration\ConfigurationChecker.cs" />
     <Compile Include="Device\IODevices\F.cs" />
-    <Compile Include="Editor\TechObjectsImporter.cs" />
+    <Compile Include="Editor\DrawInfo.cs" />
+    <Compile Include="Editor\IEditor.cs" />
+    <Compile Include="Editor\IHelperItem.cs" />
+    <Compile Include="Editor\ITreeViewItem.cs" />
+    <Compile Include="Editor\ObjectProperty.cs" />
+    <Compile Include="Editor\ImportExport\TechObjectsImporter.cs" />
+    <Compile Include="Editor\TreeViewItem.cs" />
     <Compile Include="InterprojectExchange\DeviceComparer.cs" />
     <Compile Include="InterprojectExchange\DeviceInfo.cs" />
     <Compile Include="InterprojectExchange\DeviceSignalsInfo.cs" />
@@ -208,16 +214,16 @@
       <DependentUpon>IdleForm.cs</DependentUpon>
     </Compile>
     <Compile Include="IdleModule.cs" />
-    <Compile Include="Editor\TechObjectsExportForm.cs">
+    <Compile Include="Editor\ImportExport\TechObjectsExportForm.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Editor\TechObjectsExportForm.Designer.cs">
+    <Compile Include="Editor\ImportExport\TechObjectsExportForm.Designer.cs">
       <DependentUpon>TechObjectsExportForm.cs</DependentUpon>
     </Compile>
-    <Compile Include="Editor\TechObjectsImportForm.cs">
+    <Compile Include="Editor\ImportExport\TechObjectsImportForm.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Editor\TechObjectsImportForm.Designer.cs">
+    <Compile Include="Editor\ImportExport\TechObjectsImportForm.Designer.cs">
       <DependentUpon>TechObjectsImportForm.cs</DependentUpon>
     </Compile>
     <Compile Include="Configuration\IOReader.cs" />
@@ -268,7 +274,7 @@
     <Compile Include="Editor\EditorCtrl.Designer.cs">
       <DependentUpon>EditorCtrl.cs</DependentUpon>
     </Compile>
-    <Compile Include="Editor\TechObjectsExporter.cs" />
+    <Compile Include="Editor\ImportExport\TechObjectsExporter.cs" />
     <Compile Include="TechObject\Action.cs" />
     <Compile Include="TechObject\Action_AI_AO.cs" />
     <Compile Include="TechObject\Action_DI_DO.cs" />
@@ -307,10 +313,10 @@
     <EmbeddedResource Include="IdleForm.resx">
       <DependentUpon>IdleForm.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="Editor\TechObjectsExportForm.resx">
+    <EmbeddedResource Include="Editor\ImportExport\TechObjectsExportForm.resx">
       <DependentUpon>TechObjectsExportForm.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="Editor\TechObjectsImportForm.resx">
+    <EmbeddedResource Include="Editor\ImportExport\TechObjectsImportForm.resx">
       <DependentUpon>TechObjectsImportForm.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="InterprojectExchange\Forms\FilterForm.resx">
@@ -419,6 +425,7 @@
     <Content Include="CMD\TestProjectScript.txt">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
+    <Content Include="Editor\DescriptionFromClass.txt" />
     <Content Include="LuaInterface\KopiLua.dll">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
