+++ src/Editor/DescriptionFromClass.txt
@@ -1,0 +1,52 @@
+﻿/// Описание механизма редактирования технологических объектов.
+/// В целях упрощения данного механизма используется представление данных в 
+/// виде дерева. Каждый узел дерева связан с объектом, реализующий интерфейс 
+/// ITreeViewItem, в котором определены методы (свойства), реализующие 
+/// отображение, редактирование, удаление, добавление, копирование, вставку.
+/// Класс Editor реализует данную функциональность 
+/// (Editor.Editor.Edit(ITreeViewItem корень_дерева)). 
+/// 
+/// 1.Отображение - строковое представление в виде узла дерева. Используется
+/// свойство ITreeViewItem.Name.
+///
+/// 2.Редактирование (двойной клик левой кнопкой мышки) - редактирование 
+/// строкового значения узла дерева. Редактируются определенные узлы  
+/// (ITreeViewItem.Editable). При редактировании отображается другой текст 
+/// (ITreeViewItem.EditText). Отредактированный текст обрабатывается объектом 
+/// (ITreeViewItem.SetNewValue( string )) и обновляется отображение текста 
+/// узла (ITreeViewItem.Name).
+///
+/// 3.Удаление (клавиша Del на удаляемом элементе). Действует для
+/// определенных узлов (ITreeViewItem.Deletable). Реализуется методом
+/// bool ITreeViewItem.Delete( object child ). Вызывается для родительского 
+/// узла. Если он возвращает true, данный узел удаляется. Потом обновляется
+/// отображение всех дочерних узлов (ITreeViewItem.Name).
+///
+/// 4.Добавление (клавиша Ins на узле, куда хотим вставить элемент). Действует 
+/// для определенных узлов (ITreeViewItem.Insertable). Реализуется методом
+/// ITreeViewItem ITreeViewItem.Insert(). Потом добавляется новый узел 
+/// дерева для вставленного элемента.
+/// 
+/// 5. Копирование (Ctrl + C). Действует для определенных узлов 
+/// (ITreeViewItem.Copyable). Реализуется методом
+/// ITreeViewItem ITreeViewItem.Copy(). Копируемый узел сохраняется в буферной
+/// переменной.
+/// 
+/// 6. Вставка (Ctrl + V). Действует для определенных узлов 
+/// (ITreeViewItem.InsertableCopy). Реализуется методом
+/// ITreeViewItem ITreeViewItem.InsertCopy( object obj ). Узел из буферной
+/// переменной вставляется в дерево.
+/// 
+/// 7. Отображение на странице Eplan'a - подсветка устройств.
+/// Действует для определенных узлов. Используется свойство 
+/// ITreeViewItem.IsDrawOnEplanPage.
+/// 
+/// 8. Получение списка устройств для отображения на странице Eplan'a - 
+/// подсветка устройств (открываемые устройства - зеленый прямоугольник на
+/// заднем фоне, закрываемые - красный, нижние седла - нижняя половина зеленого
+/// прямоугольника, верхние седла - верхняя половина зеленого прямоугольника).
+/// Действует для определенных узлов. Используется свойство 
+/// ITreeViewItem.GetObjectToDrawOnEplanPage.
+/// <summary>    
+/// Пространство имен редактора технологических объектов.
+/// </summary>
