+++ src/Editor/IEditor.cs
@@ -1,0 +1,12 @@
+﻿namespace Editor
+{
+    /// <summary>    
+    /// Интерфейс редактора.
+    /// </summary>
+    public interface IEditor
+    {
+        string Edit(ITreeViewItem data);
+        bool IsShown();
+        void CloseEditor();
+    }
+}
