+++ src/Editor/IHelperItem.cs
@@ -1,0 +1,10 @@
+﻿namespace Editor
+{
+    /// <summary>
+    /// Интерфейс для получения ссылки на ресурс справки по элементу
+    /// </summary>
+    public interface IHelperItem
+    {
+        string GetLinkToHelpPage();
+    }
+}
