+++ src/Editor/ImportExport/TechObjectsExporter.cs
